+++ src/components/category/coffee/lastList.js
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const lastList = () => {
+    return (
+        <>
+        <h3> 이전에 조회한 목록 </h3>
+        <hr style={{ border: "2px solid #808080", width: '100%' }} />
+        <div
+            id='lastdetail'
+            style={{
+                width: '800px',
+                height: '100px',
+                position: 'static',
+            }}
+        ></div>
+        </>
+    );
+};
+
+export default lastList;
